* feature/stop_ban_spam:
  Improve blocked user mail, add check for teams and projects persistence
  fix mail menu
  Improve blocked email body
  activate email added
  Fix mail variables
  Fix failed check of user updation
  Fix and inprove blocked user email body
  Update User Notifications Creator for banned action
  add email about user block
  fix users tests
  Prepare fix if spam on user block

+++ app/contexts/users/block_context.rb
@@ -2,14 +2,13 @@ module Users
   class BlockContext < Users::BaseContext
     def execute
       User.transaction do
-        user.block
-
-        # Remove user from all teams
-        user.user_teams.find_each do |team|
-          Gitlab::UserTeamManager.remove_member_from_team(team, user)
+        if user.block
+          # Remove user from all teams
+          user.user_teams.find_each do |team|
+            Gitlab::UserTeamManager.remove_member_from_team(team, user)
+          end
+          UsersProject.with_user(user).destroy_all
         end
-
-        UsersProject.with_user(user).destroy_all
       end
 
       receive_delayed_notifications
